overlay: Don't allow disabling of 4g calling/VoLTE
We apear to have issues with some carriers when VoLTE/4g calling features are supported but switched off because of which the modem will go into crash cycles upon reboot.

Don't allow switching off these features. The menu item is still displayed but greyed out, so still providing information on whether it is enabled.